Try1

+++ src/main/java/com/example/finaltask/model/dto/AdsDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class AdsDTO {
     private Integer author;
     private String image;
-    private Long pk;
+    private Integer pk;
     private Integer price;
     private String title;
 }
